The structures mod lives again! This is an early and very experimental implementation of the new distributed mapgen. Description:
Cities are generated based on a virtual grid, currently defaulted to 200 nodes. Each virtual cube is activated when the first mapblock in its radius is generated. When the cube is activated, a group is chosen based on the information we can gather from that first mapblock. If successful and the cube gets a group, a list of all structures is generated for that cube. Each on_generate function will only spawn those buildings within its area.

This approach should allow cities of huge sizes to be generated much more easily. Most importantly, it heavily reduces lag during world generation when a city is planned!